requirements.txt
python-binance
python-telegram-bot
tensorflow
pandas
numpy
matplotlib